Refactor package and class locations
