+++ components/org.wso2.carbon.siddhi.editor.core/src/main/java/org/wso2/carbon/siddhi/editor/core/util/Constants.java
@@ -53,9 +53,10 @@ public class Constants {
     public static final String DEPLOYMENT_PASSWORD = "password";
     public static final String SIDDHI_FILE_LIST = "siddhiFileList";
     public static final String SERVER_LIST = "serverList";
-    public static final String XML_EVENT = "xml";
-    public static final String JSON_EVENT = "json";
-    public static final String TEXT_EVENT = "text";
+    public static final String TYPE_XML = "xml";
+    public static final String TYPE_JSON = "json";
+    public static final String TYPE_TEXT = "text";
+    public static final String TYPE_CSV = "csv";
     public static final String SAMPLE_EVENTS_PARENT_TAG = "events";
     public static final String SAMPLE_EVENT_PARENT_TAG = "event";
     public static final String ATTR_TYPE_FLOAT = "float";
@@ -81,6 +82,7 @@ public class Constants {
     static final String SOURCEMAP = "SourceMap";
     static final String SINKMAP = "SinkMap";
     static final String STORE = "Store";
+    public static final String ERROR = "error";
     static final Map<String, Class<?>> SUPER_CLASS_MAP;
     static final Map<String, String> PACKAGE_NAME_MAP;
