+++ components/org.wso2.carbon.siddhi.editor.core/src/main/java/org/wso2/carbon/siddhi/editor/core/util/metainforetriever/beans/JSONConfig.java
@@ -24,22 +24,18 @@ public class JSONConfig {
     private String enclosingElement;
 
     public String getFilePath() {
-
         return filePath;
     }
 
     public void setFilePath(String filePath) {
-
         this.filePath = filePath;
     }
 
     public String getEnclosingElement() {
-
         return enclosingElement;
     }
 
     public void setEnclosingElement(String enclosingElement) {
-
         this.enclosingElement = enclosingElement;
     }
 }
